feat(core): refine employee import review workspace
